requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.8.2
boto3==1.11.9
botocore==1.14.9
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
editdistance==0.5.3
idna==2.8
jmespath==0.9.4
numpy==1.18.1
pandas==0.25.3
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
requests==2.22.0
s3transfer==0.3.2
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.8
